+++ QEfficient/utils/__init__.py
@@ -4,9 +4,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 # -----------------------------------------------------------------------------
-from typing import Union
-
-from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 
 from QEfficient.utils._utils import (  # noqa: F401
     get_qpc_dir_name_infer,
@@ -14,29 +11,7 @@ from QEfficient.utils._utils import (  # noqa: F401
     load_hf_tokenizer,
     login_and_download_hf_lm,
     onnx_exists,
+    padding_check_and_fix,
     qpc_exists,
 )
-from QEfficient.utils.logging_utils import logger
 
-
-def padding_check_and_fix(
-    tokenizer: Union[PreTrainedTokenizer, PreTrainedTokenizerFast]
-):
-    """
-    Checks Tokenizer paddding side and pad_token_id viability.
-    --------
-    
-    Tokeinizer: Union[PreTrainedTokenizer, PreTrainedTokenizerFast]. Pass model tokenizer to check and fix.
-    """
-    
-    if tokenizer.padding_side != "left":
-        logger.warning("Please use padding_side='left' while initializing the tokenizer")
-        tokenizer.padding_side = "left"
-    
-    if tokenizer.pad_token_id is None:
-        # If Pad token is out of range of vocab size
-        if tokenizer.eos_token_id < tokenizer.vocab_size:
-            tokenizer.pad_token_id = tokenizer.eos_token_id
-        else:
-            tokenizer.pad_token_id = tokenizer.vocab_size - 1
-
